UDP socket interface and Client/Server
I finished implementing the UDP socket interface. I had to fix a few things
with the socket itself. It was not reading the client ip address correctly, as
I was passing the pointer incorrectly. I have fixed that. I've also finished
implementing the write method and it sends the data.

I've added a logs file as per project spec.