Adding Chord Client system
This is the basic interface with the chord system.
We will be able to get our own value and lookup the value of someone around us
based on a key or a value.

The system will use a heartbeat and pulse to detect when its neighbors have
died. This will be done on a separate thread every so often. If a node fails to
emit three or so heartbeats, we will assume the node to be dead.