Using Unordered Map
The hash table uses unordered maps from the STL. This should give us better
support.

Added semantic meaning to comparisons for the hashes. This means that we can
remove the ChordKey class as it adds nothing on top of Hash.

Added remove to the hash table, which deletes the item.