UID passing and neighbors
Now that communications are working correctly, we need to start building the
chord ring.

Chord takes an extra parameter, uid (unique identifier) which is hashed to
determine the relative position of a node in the ring. This can be done very
easily.

We may need to copy the node's ip and port number into the header to propagate
that to the correct node for storage. Right now it is inaccessible because the
node at the receiving end will see its neighbor's address and not the new node.