Protocol Update
Updating protocol to carry bytes in payload.
Also optionally carries ip address, port number, and forwarding boolean.

When the forwarding boolean is set, we read the ip and port number from the
packet rather than from the ip header because it was sent through a peer and we
don't want the peer's address, we want the originating node's address.