Raw Hash
For transporting the hash contents, we don't want the string, we want to copy
the contents of the payload directly into the hash buffer.

The constructor takes a string and optionally a boolean. We assume that the
input string is not hashed and will hash it unless the boolean is set, in which
case we copy the contents of the string into the hash buffer.

We also output the raw hash buffer using the raw() method.