Join Request
This is one attempt at a join request. I'm not sure that it is working
correctly, or how I want it to at this point. I may be doing some major changes
to the protocol to enable responses. At this point everything is request based.

Possible future protocol buffer may look like:

Packet:
	string ip
	i32 port
	bytes id (hash bytes)
	boolean pass
	type (request, response, ping):

request:
	enum Types
		JOIN
		DROP
		GET
		SET
		SYNC
		UPDATE

response:
	enum Types
	{
		GOOD
		BAD
	}

ping:
	bool pong