Join Cleanup part 2
I've removed a bunch of duplicate code for handling a forwarded join request.
It is possible that I could have the request handler immediately detect if the
request has been forwarded and simply place the correct ip/port number into the
function.

I've also removed a lot of the print statements, making the output easier to
read and debug.