WIP | Node failure handling
A node that is not the originator node can fail and the chord will attempt to
repair itself after a three missed heart beats.

We can use the heart beats to determine who our predecessor is because it will
be coming from them. Then we can correctly follow the Chord algorithm again.

Todo:
	If the original node dies, the whole system gets messed up because some
	node will be unable to fix the ring. From there, even if the original
	node comes back up, the ring will be broken and unable to repair
	without resetting everyone else.

	If a node dies, then comes back up, we may either want to ping it more
	often, or less often. If the issue is that there is congestion and the
	packets are getting dropped, then less often would be better.