Basic Logging header
I've added a super basic logging header. It takes a string and writes it to a
desired output file. All log messages have the date and time placed at the
beginning of the line followed by a colon.

Format is dd/mm/yyyy hh:mm:ss am|pm: <message>

It is using 24-hour local time to the server.

Example:
29/06/2015 08:17:02 am: Incorrect arguments: 2 provided, 3 expected, 5 optional