Adding Extra Debug information
I've added a way of dumping the current state of a node to a dump file.
This has exposed an infinite loop between nodes. It appears that the hash is
updated, but the ip address and port number are not updated in the successor
node. We'll have to check this out a bit more.