Relaxing Requirements
I've relaxed the requirements of the system. We are no longer strictly working
with the Chord implementation, but will use some of the ideas from Chord for
the implementation of the system.

I will be switching from the Google Protocol buffers because they seem to be
leaving holes in my data, especially surrounding the unknown length fields (the
id and ip address). The length of transmission shouldn't be an issue as the
largest packets I've seen have been 69 bytes, though usually less.

I will be trying to use Capt'n Proto. If that doesn't work, I'll be switching
to just using JSON for passing the requests.