Fixing Socket
When reading from the socket, no size was provided to the std::string
constructor, causing it to use NULL termination method to determine the length.
As the contents of the buffer was binary data, nulls were possible within the
buffer that were not the end.

The fix specifies the number of bytes, making the outgoing string more
reliable.