Cleaning up and adding to protobuf
I've cleaned up the output of the program to make it easier to understand what
is happening. I've also removed a fair amount of commented out code, though
most of it was commented out output.

I've also fixed some things with the port numbers. If a node is behind NAT,
which we are not handling, it wouldn't be able to determine its own port, so
for foresight, we read the port from the socket unless the packet has been
forwarded.

I've also added a few more Request types, which are actually responses.
	BAD is for when the lookup doesn't exist.
	RES is the response for when it does exist.