Working DNS Resolution
By golly, it is working.
There is still a lot of work to be done before it is at all stable or usable
though.

Currently there can be one client for each server. This is because of the
method used for passing the response back to the requester.

The future will use hash maps to map a request to many waiting requesters,
allowing for fewer requests to be sent out. If many requesters make a request
for the same resource, then we can store all the requesters in a list and give
assign all of them the first returned address that matches the request.

Next, some cleanup though.
I'll be using the logs to show who asked for what at what time. That way the
application will remain looking clean through the operation.