Adding Partial Predecessor Stabilization code
To model after Chord, we are using the same predecessor stabilization algorithm
to attempt to handle node removals mode gracefully and hopefully even allow the
system to continue functioning in the event of node failure.

We will also have to poll our predecessor to see if it is still alive.