Determining Predecessor position
We send the ip, port, and id of our predecessor back on a sync request. This
way a requester can update their successor or confirm that their successor is
who they think it is. This way if something is incorrect, it will get fixed.
This also allows a node to attempt to set its successor to any node it chooses
and eventually will find the correct successor.

If we have a finger table, we can attempt to make our successor the closet
finger that will connect. It will then redirect us to its predecessor if that
predecessor is still alive, and we will eventually find the successor we are
looking for.