Correcting SYNC documentation in chord_message.proto
Since we are using the SYNC in the stabilization of the chord ring, we need to
change some of the definitions for the sync field.

We need to pass the predecessor hash, ip, and port back to the node that
believes it is our predecessor. That way it can check the hash and see if it is
actually our predecessor. If it isn't, it will be able to change it's successor
value and try that node, which will then pass its predecessor id, ip, and port.
This allows nodes to eventually find the correct place in the ring if something
should get messed up after the initialization phase of the ring.