Ring repairs itself if a given node fails
The ring repairs itself eventually. Right now it is very slow due to the delay
caused by the heartbeats. If we increase the heart rate when the successor or
the predecessor haven't confirmed the death of a node, then the heart rate
should be increased to repair the system faster.