Chord Repair Down to two node
With the previous setup, once we got below two nodes, things would break
because the systems would never accept the other as their successor. The fix
appears to repair that, but it is VERY slow (minutes for the ring up to 8
nodes). It would be good to go over the changes and see if we can improve
performance.

That said, I was able to grow the ring up to 8 nodes, then shrink it back down
to two nodes. If I waited for everything to stabilize between node kills, it
regroups correctly, eventually...