Chord Rejoining Faster
I have the heart rate increase when the chord nodes start noticing that
something is wrong. This is to minimize the re-stabilization time.

I've also added a new command 'd', which outputs the same information as the
'dump' command, but instead of writing it to a file, it writes to the
console for faster debugging.