Adding Base Fabric File
Used to coordinate GENI nodes.

On the machines I was able to get the program running on, it worked
beautifully.

I'm having some issues installing some of the required packages on a few nodes.
They are unable to resolve the ubuntu archive hostnames. I may simply keep a
local copy of the library and have it push that into /usr/lib